% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBlastresults.R
\name{checkHybridisation}
\alias{checkHybridisation}
\title{Check hybridisation status of each hit}
\usage{
checkHybridisation(
  coverage,
  pident,
  var_type,
  variants,
  var.position,
  istarget3primeend = FALSE,
  min.coverage,
  min.pident,
  min.matchbp,
  marker.char
)
}
\arguments{
\item{variants}{list of variants and their positions}

\item{istarget3primeend}{Boolean value showing whether the target marker is on the 3 prime end}

\item{min.coverage}{minimum coverage as percentage}

\item{min.pident}{minimum percentage identity}

\item{min.matchbp}{Integer showing the number of expected correctly matched base pairs}

\item{marker.data}{Row list of marker data}
}
\description{
Function used to check the hybridisation status of a probe sequence
}
\author{
David Chisanga
}
