% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doStrictfiltering.R
\name{DoStrictfiltering}
\alias{DoStrictfiltering}
\title{Do strict filtering}
\usage{
DoStrictfiltering(
  probe.name,
  genome.name,
  output.path,
  dotargetfilt,
  chrom.comp,
  target.data,
  domapmarkers,
  similarity.maps,
  doldedge,
  ldmapp,
  blast.hits,
  mappings.file,
  dogeneticmap,
  geneticmap.file,
  dupmapinter.file
)
}
\arguments{
\item{probe.name}{name of the probe file being analysed}

\item{genome.name}{name of the reference genome being compared}

\item{output.path}{path were the results are saved}

\item{dotargetfilt}{whether to do target chromosome based filtering}

\item{chrom.comp}{Location of chromosome comparison pairs between old an new references}

\item{target.data}{2 column table with first column showin allele ID and second expected target chrom to preference}

\item{domapmarkers}{whether to do key linked marker based filtering}

\item{similarity.maps}{X column table where the first column is the target marker and n columns after contain a marker known to be linked to the target marker}

\item{doldedge}{whether to do linkage disequilibrium based filtering}

\item{ldmapp}{10 column table (standard .ld format) for pairwise ld scores and locations per marker}

\item{blast.hits}{Intermediate filtered blast hits file as csv}

\item{mappings.file}{Low filtered mappings file as tsv}

\item{dogeneticmap}{whether to use genetic mapping prior}

\item{geneticmap.file}{CSV path to genetic mapping prior (MarkerName, Chromosomes_mapped, etc)}

\item{dupmapinter.file}{file tracking the presence of local duplications of markers in each genome. multi column header tsv}
}
\value{
Returns an object which has been filtered using given priors to remove/keep a maximum of 1 hit (the best) per marker
}
\description{
This function is used to undertake an strict level of filtering producing output files with only one hit per marker allowed
}
\author{
James O'Dwyer
}
\keyword{blast}
\keyword{filter}
\keyword{genome,}
\keyword{strict}
