% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMarkerInformation.R
\name{getAltBP}
\alias{getAltBP}
\title{Get the alternative base for a probe ID}
\usage{
getAltBP(NT_A, NT_B, Ref_NT)
}
\arguments{
\item{NT_A}{Design cluster A nucleotide}

\item{NT_B}{Design cluster B nucleotide}

\item{Ref_NT}{Nucleotide at SNP position from reference genome}
}
\value{
Returns a character string indicating the alternative base pair
}
\description{
This function used is to get the alternative base pair from blast results
}
\author{
David Chisanga
}
\keyword{allele}
\keyword{allele,}
\keyword{alternative}
\keyword{genome,}
\keyword{reference}
