% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBlastresults.R
\name{getPositionsFromBTOP}
\alias{getPositionsFromBTOP}
\title{Function to get variant positions from blast btop}
\usage{
getPositionsFromBTOP(btop)
}
\arguments{
\item{btop}{String btop value}
}
\description{
The function is used to obtain the positions of variants
from the btop output from blast
}
\examples{
getPositionsFromBTOP("4A-40-AGC25TA5")
}
\author{
David Chisanga
}
\keyword{btop,blast,variant,position}
