% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBlastresults.R
\name{getVariantsFromBTOP}
\alias{getVariantsFromBTOP}
\title{Get Variants from BLAST BTOP}
\usage{
getVariantsFromBTOP(btop)
}
\arguments{
\item{btop}{String btop value from blast}
}
\description{
Function to separate variants within the btop string outpt by blast
}
\examples{
getVariantsFromBTOP("4A-40-AGC25TA5")
}
\author{
David Chisanga
}
\keyword{btop,variants,blast}
