% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBlastresults.R
\name{parse_btop}
\alias{parse_btop}
\title{Parse BTOP from BLAST}
\usage{
parse_btop(btop)
}
\arguments{
\item{btop}{String btop value from blast}
}
\description{
Function to separate variants within the btop string outpt by blast
}
\examples{
parse_btop("4A-40-AGC25TA5")
}
\author{
David Chisanga
}
\keyword{btop,blast}
