% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBlastresults.R
\name{processBlastResults}
\alias{processBlastResults}
\title{Process alignment results from blast}
\usage{
processBlastResults(
  blast.file,
  min.length = 40,
  path.2save.coords = "SNPcoordinates",
  extendable.site.bps = 3,
  targets,
  outfmt = c("qaccver", "saccver", "pident", "qlen", "length", "mismatch", "gapopen",
    "qstart", "qend", "sstart", "send", "evalue", "bitscore", "btop", "qseq", "sseq",
    "sstrand"),
  istarget3primeend,
  marker.char = "N",
  min.coverage = 70,
  min.pident = 90
)
}
\arguments{
\item{blast.file}{path to the blast output file}

\item{path.2save.coords}{path where SNP coordinates to get base in reference genome using samtools are to be saved}

\item{extendable.site.bps}{number of matching base pairs from the 3 prime end for a probe to be considered as extendable}

\item{outfmt}{vector of format specifiers from the supported format specifiers for 6,7 and 10 in blastn's 'outfmt' parameter}

\item{istarget3primeend}{logical value showing if target is on the 3' prime end}

\item{min.matchbp}{minimum length of HSP to be considered fully Hybridize}

\item{target}{table with target annotation}

\item{num.threads}{integer value of the number of threads passed to reading the data file}

\item{keep.duplicates}{logical value, whether to retain lines with multiple hits}
}
\value{
Returns a data frame of processed blast results
}
\description{
This function is used to process and format the blast output
}
\author{
David Chisanga
}
\keyword{alignment}
\keyword{blast,}
