% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doIntermediatefiltering.R,
%   R/doStrictfiltering.R
\name{read_mixed_csv_base}
\alias{read_mixed_csv_base}
\title{Read in file with comment lines and attach those lines to read in object}
\usage{
read_mixed_csv_base(path, sep = ",", header = TRUE, parse_kv = TRUE, ...)

read_mixed_csv_base(path, sep = ",", header = TRUE, parse_kv = TRUE, ...)
}
\arguments{
\item{path}{Character. Path to the file to read.}

\item{sep}{Field separator for the data section (default \code{","}). Use \code{"\\t"} for TSV.}

\item{header}{Logical; whether the first non-header line contains column names. Default \code{TRUE}.}

\item{parse_kv}{Logical; if \code{TRUE}, parse \verb{## key=value} lines into \code{attr(df, "header_meta")}.}

\item{...}{Additional arguments passed to \code{utils::read.csv()}.}
}
\value{
Returns an object like a read.table or read.csv function with comment.char applied but with the comment lines attached to the attributes of the object so they are not lost

Returns an object like a read.table or read.csv function with comment.char applied but with the comment lines attached to the attributes of the object so they are not lost
}
\description{
This function is used to read in csv files in the Brioche pipelines to preserve their appended headers of filtering information

This function is used to read in csv files in the Brioche pipelines to preserve their appended headers of filtering information
}
\author{
James O'Dwyer
}
\keyword{comment}
\keyword{csv,}
\keyword{genome,}
\keyword{lines}
