% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doIntermediatefiltering.R,
%   R/doStrictfiltering.R
\name{sanitize_rectangular}
\alias{sanitize_rectangular}
\title{write out file with comment lines}
\usage{
sanitize_rectangular(x)

sanitize_rectangular(x)
}
\arguments{
\item{x}{Data-frame-like object to sanitise.}
}
\value{
A \code{data.frame} with list/data.frame columns collapsed to one string per row.

A \code{data.frame} with list/data.frame columns collapsed to one string per row.
}
\description{
This function is used to sanitise brioche blast hit objects to ensure each column or row has the same length

This function is used to sanitise brioche blast hit objects to ensure each column or row has the same length
}
\author{
James O'Dwyer
}
\keyword{genome,}
\keyword{rectangular}
\keyword{sanitize,}
