% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMarkerInformation.R
\name{addSNPdetails}
\alias{addSNPdetails}
\title{Add SNP details to the blast output}
\usage{
addSNPdetails(
  blast.path,
  reference.bases,
  probe.name,
  genome.name,
  output.path = "Mapping_results",
  pident.filter = 90,
  coverage.filter = 50,
  is3prime
)
}
\arguments{
\item{blast.path}{Dataframe or path to blast results returned by the
\code{processBlastResults()} function (RDS file).}

\item{reference.bases}{Path to the reference bases obtained from samtools.}

\item{probe.name}{Name of the probe file being analysed.}

\item{genome.name}{Name of the reference genome being compared.}

\item{output.path}{Path where the results are saved.}

\item{pident.filter}{Double value to filter hits with small percentage identity.}

\item{coverage.filter}{Double value to filter hits with small coverage.}

\item{is3prime}{Logical. Does the SNP exist as 3' as part of SNP chip marker mapping.}
}
\value{
Returns a dataframe with the updated blast results. The returned object
has attributes \code{is3prime}
}
\description{
This function is used to add more details to the blast output
returned by the \code{processBlastResults()} function
}
\author{
David Chisanga
}
\keyword{allele}
\keyword{allele,}
\keyword{alternative}
\keyword{genome,}
\keyword{reference}
